+++ lua/obsidian/commands/init.lua
@@ -300,9 +300,9 @@ M.register("bookmarks", { nargs = 0 })
 
 M.register("backlinks", { nargs = 0, note_action = true })
 
-M.register("incoming_links", { nargs = 0, note_action = true })
+M.register("find_incoming", { nargs = 0, note_action = true })
 
-M.register("outgoing_links", { nargs = 0, note_action = true })
+M.register("find_outgoing", { nargs = 0, note_action = true })
 
 M.register("template", { nargs = "?", note_action = true })